[charts] DRY owned REST APIs
